requirements.txt
boto3==1.34.19
python-dotenv==1.0.0
requests==2.31.0
botocore==1.34.19
